Split RepomdFileCollector from the yum downloader code
